+ Added ScriptModelBlockbench.java for using the engine Blockbench which uses different x/y/z calculations and rotations + Added an abstract base class for the model handling
