+++ src/main/java/dev/adventurecraft/awakening/mixin/client/render/MixinWorldEventRenderer.java
@@ -13,6 +13,7 @@ import dev.adventurecraft.awakening.extension.entity.ExLivingEntity;
 import dev.adventurecraft.awakening.extension.world.chunk.ExChunkCache;
 import dev.adventurecraft.awakening.common.AC_Particle;
 import dev.adventurecraft.awakening.script.ScriptModel;
+import dev.adventurecraft.awakening.script.ScriptModelBase;
 import net.minecraft.block.Block;
 import net.minecraft.class_66;
 import net.minecraft.client.Minecraft;
@@ -913,7 +914,7 @@ public abstract class MixinWorldEventRenderer implements ExWorldEventRenderer {
     private void renderScriptModels(Vec3d var1, CameraView var2, float var3, CallbackInfo ci) {
         GL11.glPushMatrix();
         GL11.glTranslated(-EntityRenderDispatcher.field_2490, -EntityRenderDispatcher.field_2491, -EntityRenderDispatcher.field_2492);
-        ScriptModel.renderAll(var3);
+        ScriptModelBase.renderAll(var3);
         GL11.glPopMatrix();
     }
