+++ src/main/java/dev/adventurecraft/awakening/script/ScriptModelBlockbench.java
@@ -1,0 +1,19 @@
+package dev.adventurecraft.awakening.script;
+
+public class ScriptModelBlockbench extends ScriptModelBase {
+
+    @Override
+    protected void transform(float deltaTime) {
+
+    }
+
+    @Override
+    protected void render(float var1) {
+
+    }
+
+    @Override
+    protected void update() {
+
+    }
+}
